Make Python 3 requirement more explicit
If you try to install tootstream under Python 2, it will look like it
installed but then immediately throw an exception upon running due to
non-ASCII characters in the source code. Specify Python 3 earlier in the
README and specify `python_requires` in `setup.py` so it checks the
Python version in the installation process.